tweak bio to use local fs image, copy
