+++ src/pages/blog.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link, graphql } from 'gatsby';
 
-import Bio from '../components/bio';
+import Bio from '../components/Bio';
 import Page from '../components/Page';
 import SEO from '../components/seo';
